機能追加: DocRowコンポーネントに変更状態を表示するChangeBarコンポーネントを実装し、表示ロジックを簡素化
